confirm message before delete post

+++ src/components/posts_show.js
@@ -11,10 +11,12 @@ class PostsShow extends Component {
     }
 
     onDeleteClick() {
-        const { id } = this.props.match.params;
-        this.props.deletePost(id, () => {
-            this.props.history.push('/');
-        });
+        if(confirm('Delete the item?')) {
+            const { id } = this.props.match.params;
+            this.props.deletePost(id, () => {
+                this.props.history.push('/');
+            });
+        }
     }
 
     render() {
